Update setup_homedir and setup_groupdir
Needed to update these scripts to fix a few things. For the homedir
script, we needed to ensure we were treating domains correctly, so
domain enforcement was added and noted in how you need to call the
script.  The homedir script was also updated to enforce a lowercase
letter requirement on all usernames.  If zfs datasets got made with
upper case letters in them, they wouldn't match when a user logged
in and the user would get a folder in the main zfs dataset.  The
groupdir script was also updated to enforce lower case for consistency
sake.