Minor update to synthetik report and setup_homedir
Sythetic report had an extra debug line added to check on some
output I was seeing that I was sure if it was anomolous or not. It
appears to be working correctly.  The setup_homedir update checks
incoming usernames to make sure they are not over 20 characters
long.  This will avoid an issue were the script can create a zfs
dataset for a users home directory and then that directory is not
used by the system when the user logs in because it doesn't match
the username that comes back from sssd/ad/etc.  It reminds script
users to use the windows pre-2000 style, 20 character usernames in
the script.