Updating synthetik report script
The script now emails users, and has a robust argument system that
allows users to easily turn on or off different features of the
script.  This means you can specify whether or not to email, how
much information to print to screen, what emails to send to, and
more all from the command line.  The script assumes that the
localhost has an smtp relay to send emails with.